authUser fix
