--- src/types/AuthUser.ts
@@ -1,19 +1,0 @@
-export interface LoginRequest {
-
-  username: string;
-
-  password: string;
-
-}
-
-
-
-export interface AuthUser {
-
-  token: string;
-
-  username: string;
-
-  role: string;
-
-}
